Setted up the project with maven.
made a new project that works with maven and can create .jar files
