--- README.md
@@ -1,2 +1,0 @@
-# tda367-owls
-Repo för Objektorienterat programmeringsprojekt
