--- src/main/java/edu/chalmers/projecttemplate/controller/ProjectController.java
@@ -1,20 +1,0 @@
-package edu.chalmers.projecttemplate.controller;
-
-import edu.chalmers.projecttemplate.model.Project;
-import edu.chalmers.projecttemplate.view.ProjectView;
-import javafx.fxml.Initializable;
-
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.net.URL;
-import java.util.ResourceBundle;
-
-public class ProjectController implements Initializable {
-    private Project project;
-    private ProjectView projectView;
-
-    @Override
-    public void initialize(URL url, ResourceBundle rb) {
-
-    }
-}
