--- src/main/java/edu/chalmers/projecttemplate/model/Project.java
@@ -1,16 +1,0 @@
-package edu.chalmers.projecttemplate.model;
-
-public class Project {
-    public static final String PROJECT_WINDOW_TEXT = "ProjectTemplate";
-    public static final String PROJECT_BUTTON_TEXT = "Press me!";
-
-    private int presses;
-
-    public int getPresses() {
-        return presses;
-    }
-
-    public void incrementPresses() {
-        this.presses++;
-    }
-}
