--- src/main/java/edu/chalmers/projecttemplate/view/ProjectView.java
@@ -1,10 +1,0 @@
-package edu.chalmers.projecttemplate.view;
-
-import edu.chalmers.projecttemplate.model.Project;
-
-public class ProjectView {
-
-    public ProjectView(Project project) {
-
-    }
-}
