--- src/test/java/edu/chalmers/projecttemplate/model/ProjectTest.java
@@ -1,19 +1,0 @@
-package edu.chalmers.projecttemplate.model;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-public class ProjectTest {
-	private static final int NUM_INCREMENTATIONS = 128;
-
-	@Test
-	public void testIncrementResult() {
-		final Project project = new Project();
-
-		for (int i = 0; i < NUM_INCREMENTATIONS; i++) {
-			project.incrementPresses();
-		}
-
-		Assert.assertEquals(NUM_INCREMENTATIONS, project.getPresses());
-	}
-}
