+++ storIT/src/main/java/storagesystem/controller/ProjectController.java
@@ -1,0 +1,16 @@
+package storagesystem.controller;
+
+import javafx.fxml.Initializable;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class ProjectController implements Initializable {
+    //private Project project;
+    //private ProjectView projectView;
+
+
+    public void initialize(URL url, ResourceBundle rb) {
+
+    }
+}
